Removes old files from previous Xcode
